rename "roll-table.ts" + "table-result.ts"
