+++ src/module/roll-table/oracle-table-result.ts
